+++ chat-widget/src/common/telemetry/TelemetryConstants.ts
@@ -35,6 +35,7 @@ export enum BroadcastEvent {
     RedirectPageRequest = "RedirectPageRequest",
     StartChatSkippingChatButtonRendering = "StartChatSkippingChatButtonRendering",
     StartUnauthenticatedReconnectChat = "StartUnauthenticatedReconnectChat",
+    SetCustomContext = "SetCustomContext"
 }
 
 // Events being logged
@@ -139,7 +140,8 @@ export enum TelemetryEvent {
     ReconnectChatMinimize = "ReconnectChatMinimize",
     
     MessageSent = "MessageSent",
-    MessageReceived = "MessageReceived"
+    MessageReceived = "MessageReceived",
+    CustomContextReceived = "CustomContextReceived"
 }
 
 export interface TelemetryInput {
@@ -187,6 +189,9 @@ export class TelemetryConstants {
         case TelemetryEvent.PreChatSurveyStartChatMethodFailed:
         case TelemetryEvent.HeaderCloseButtonClicked:
         case TelemetryEvent.HeaderMinimizeButtonClicked:
+        case TelemetryEvent.MessageSent:
+        case TelemetryEvent.MessageReceived:
+        case TelemetryEvent.CustomContextReceived:
             return ScenarioType.ACTIONS;
 
         case TelemetryEvent.StartChatSDKCall: